pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "releasy"
version = "0.1.0"
description = "CLI tool for managing fork rebases, feature branches, and release construction"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "pygithub>=2.1",
    "rich>=13.0",
]

[project.scripts]
releasy = "releasy.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/releasy"]

[tool.ruff]
line-length = 100
target-version = "py39"
